+++ CountlyCommon.h
@@ -22,6 +22,7 @@
 #import "CountlyNotificationService.h"
 #import "CountlyConsentManager.h"
 #import "CountlyLocationManager.h"
+#import "CountlyRemoteConfig.h"
 
 #if DEBUG
 #define COUNTLY_LOG(fmt, ...) CountlyInternalLog(fmt, ##__VA_ARGS__)
@@ -58,6 +59,15 @@
 extern NSString* const kCountlySDKVersion;
 extern NSString* const kCountlySDKName;
 
+extern NSString* const kCountlyErrorDomain;
+
+NS_ERROR_ENUM(kCountlyErrorDomain)
+{
+    CLYErrorFeedbackWidgetNotAvailable = 10001,
+    CLYErrorFeedbackWidgetNotTargetedForDevice = 10002,
+    CLYErrorRemoteConfigGeneralAPIError = 10011,
+};
+
 @interface CountlyCommon : NSObject
 
 @property (nonatomic) BOOL hasStarted;
@@ -80,6 +90,7 @@ void CountlyInternalLog(NSString *format, ...) NS_FORMAT_FUNCTION(1,2);
 
 #if (TARGET_OS_IOS || TARGET_OS_TV)
 - (UIViewController *)topViewController;
+- (void)tryPresentingViewController:(UIViewController *)viewController;
 #endif
 
 - (void)startAppleWatchMatching;